requirements.txt
asgiref==3.8.1
Django==5.1.3
django-crispy-forms==2.3
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.2
